requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0
scipy>=1.10.0
